Issue 10: fix bug with OverallStatsNumbers; change graphEntries to Double; exclude unfinished events

+++ src/main/java/property/abolish/archery/http/model/responses/OverallStatsGraphResponse.java
@@ -3,9 +3,9 @@ package property.abolish.archery.http.model.responses;
 import java.util.List;
 
 public class OverallStatsGraphResponse {
-    public OverallStatsGraphResponse(List<Integer> graphEntries) {
+    public OverallStatsGraphResponse(List<Double> graphEntries) {
         this.graphEntries = graphEntries;
     }
 
-    public List<Integer> graphEntries;
+    public List<Double> graphEntries;
 }
